backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
mutagen==1.47.0
pyacoustid==1.3.0
httpx==0.27.0
websockets==13.0
python-multipart==0.0.9
musicbrainzngs==0.7.1
numpy==2.2.3
psycopg2-binary==2.9.10
